[Java] Use a separate UnsafeBuffer to compute CRC in the RecordingWriter
We can't mutate the underlying log buffer when computing CRC checksums in the recording.
Doing this leads to the buffer not being identified correctly in the DataPacketDispatcher.
Therefore computing CRC should happen on the copy of the buffer which can be freely mutated
and then persisted to the file system.

To avoid extra costs the copy happens iff the CRC is enabled for the recording and the data
frames are being processed.